HT-2463: transaction support, retry support
Transactions should help address some kinds of race conditions where
clusters are being merged (merge_many) or split (reclusterer). Don't
know if it's actually necessary, but worth seeing if we end up with
something consistent with them turned on, and if it doesn't affect
performance too badly.

This will break travis, need to wait for mongo test to be up and then
initiate the replica set

This also addresses some things that can happen when a document is
deleted out from under a process -- it will try to re-insert it and get
a duplicate key error. The upserts are rather ugly, but mongoid doesn't
expose a way to indicate if an update actually affected any documents.

Also adds scripts to set up replica set both for dev & testing;
.travis.yml just calls the bin script so it's easy to replicate in dev
what we are doing in travis.