Naive approach to computing hscores
HtItemOverlap uses cluster_overlap to provide a list of unique members that hold an item. bin/compile_hscores leverages this to compute a frequency table of member=> { num_members : num_matching_items } and a total hscore.
Actual report format is TBD; currently a JSON dump of the totals table.

The set of members that hold a particular item is generated when the hscores are being compiled, which necessitates a retrieval of the holdings records. This is suboptimal as the set of members could be retained in the HtItem, if we ensure it is updated when appropriate.