Clean up use of HTMembers in tests
There were some issues with reuse of the database connection leading to
odd errors elsewhere in the test suite. This change tries to ensure that
the specs leave everything in a clean state for any other tests that
might rely on the HTMembers database connection, and mocks HTMember data
in spec_helper.rb so that by default we don't hit the database.