Add small but complete set of test data for SPMs
This adds a small set of inter-related data, with fake
hathifile, OCLC concordance, and a script to generate
fake holdings with characteristics useful for integration
testing. The latter exercise the concordance, shared and
unshared items, items not in the hathifiles, repeated
items, etc.  See `generate_mono.rb` for comments
describing all the effects.