Try to prevent state from leaking between tests
This fixes issues with the logger getting set to STDERR (thus polluting
test output) as well as with unexpected external HTTP calls.

- Consistently wrap stuff in bin scripts in "main" method to avoid side
effects when loading in specs (in particular to avoid changing stuff in
Services). This then leads to rubocop complaints about stuff in main,
which is probably reasonable, but also not really caused by this commit
- hence added to .rubocop_todo.yml.

- Avoid use of leaky constant BATCH_SIZE

- Avoid things that happened when autoscrub classes or specs were loaded
rather than when a specific spec is run or a class is instantiated -- in
particular fetching the max OCN

- Use webmock to stub logging calls to Slack and to the pushgateway