Output to file for reports
The logic to generate a filename is similar between many of the reports
and could probably be extracted.

The list of output paths is getting pretty long and messy, and we should
think about what organization we really want there.

Also, disables exposing port for pushgateway - not really needed and it
was fighting with something else unrelated in a different project.