Refactor sidekiq jobs
Add Jobs::Common as a common interface between thor, the sidekiq jobs queue,
and the logic. Thor doesn't produce keyword options by default, and
sidekiq can't handle them anyway; Jobs::Common marshals the job
parameters, constructs an instance of the actual class we want, and runs
the job. Not all of the jobs in Jobs are amenable to that pattern yet,
but it does reduce the amount of boilerplate when we can use it. The
jobs using FileLoader are almost amenable to having a similar pattern.

This also moves the integration tests that were in sidekiq_job to a
separate integration test spec that uses phctl directly to test
everything end-to-end.