Add option to phctl to run inline
Having a global option for thor is a little bit clunky, but using the
"class_option" on the parent class and then handling the option in a
single place that the subcommands inherit from seems to work. If there
is a better way to do this with Thor, I couldn't find it.