DEV-505: Only use resolutions & htitems when updating OCNs
This change ensures that when OCNs in a cluster change, we only use the
OCNs for the resolutions and the htitems. In normal usage, holdings
should never end up in the "wrong" cluster. However, a small number of
shared print commitments had their OCNs changed without moving clusters,
leading to inconsistent clusters. This update mitigates the effect of
that. DEV-504 will address the underlying issue.