DEV-1528/DEV-1529: optimize holdings load
When loading holdings, for the time being we just need to push rows into
mariadb as fast as possible. There aren't any existing holdings we need
to worry about, and we don't need to check or add rows in cluster_ocns
-- we only need to worry about clusters with OCNs attested in hathifiles
or OCLC resolutions.

When we want to update holdings, I suspect we'll want to delete existing
holdings in bulk and then add the new ones. If we aren't just deleting
existing holdings, we may want to add a primary key on UUID and either
update/ignore on duplicate key which gives us idempotence to some
extent; if we don't need to do that, we may be able to drop the UUID
column entirely.

Next step is looking at having FileLoader construct larger batches for
Holdings.