Provide cluster to HtItem, Holding on construction
If we start with e.g. cluster.for_ocns(...) and then reach through for
holdings and ht items, if we then ask for the cluster again this will do
more sql queries. This way we can provide the cluster directly when
constructing the htitem/holding so it doesn't need to re-query.