ETT-408: report for analyzing whether depositor holds htitem
In the future we may want to consider an option in Cluster to load
counts for one organization rather than for all organizations. I don't
think we need it now since this report isn't time-sensitive and
shouldn't take any longer than the cost report anyway. But we might want
to consider it if we start needing to use the item_access holdings API a
lot more, since right now that looks at more rows than it needs.