ETT-1178: option to skip type checking
This is desirable when for example we are loading a new members'
holdings, one type passes validation, but another doesn't; if we then
fix the one that failed, we get the complaint about type mismatch (since
only one type of data was loaded before).

This is also potentially an issue when formats do change and we want to
manually back up, delete, and then reload. Again, when loading new data, we can
run into this problem if some passes the validation & we load it but
other data doesn't.

Longer-term, we should think about:

* by default, loading data only if *all* submitted files are valid
* being able to specify a particular file to scrub & load when needed
* automating the reload process when formats change (ETT-253)