Include icus items for deposit holdings analysis
Because the deposit holdings analysis is primarily for resource sharing
which is (for now) a US-based service, we wanted to include icus
material in that analysis.

This also exposes functionality for the mock solr response to include
different records or queries, which could be useful for certain kinds of
testing.